# internal/handlers/auth.handler.go
package handlers

import (
	"net/http"
	"time"

	"github.com/cristian-yw/Weekly10/internal/middleware"
	"github.com/cristian-yw/Weekly10/internal/models"
	"github.com/cristian-yw/Weekly10/internal/repository"
	"github.com/gin-gonic/gin"
	"github.com/redis/go-redis/v9"
	"golang.org/x/crypto/bcrypt"
)

type AuthHandler struct {
	ar          *repository.AuthRepository
	RedisClient *redis.Client
}

func NewAuthHandler(ar *repository.AuthRepository, rdb *redis.Client) *AuthHandler {
	return &AuthHandler{ar: ar, RedisClient: rdb}
}

// @Summary Register new user
// @Description Create a new user account
// @Tags Auth
// @Accept json
// @Produce json
// @Param request body models.RegisterRequest true "Register info"
// @Success 201 {object} models.SuccessMessage
// @Failure 400 {object} models.ErrorResponse
// @Failure 409 {object} models.ErrorResponse
// @Router /auth/register [post]
func (h *AuthHandler) Register(c *gin.Context) {
	var req models.RegisterRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, models.ErrorResponse{Error: "Invalid input"})
		return
	}

	hash, err := bcrypt.GenerateFromPassword([]byte(req.Password), bcrypt.DefaultCost)
	if err != nil {
		c.JSON(http.StatusInternalServerError, models.ErrorResponse{Error: "Failed to hash password"})
		return
	}

	if err := h.ar.RegisterUser(req.Email, string(hash)); err != nil {
		c.JSON(http.StatusConflict, models.ErrorResponse{Error: "Email already exists"})
		return
	}

	c.JSON(http.StatusCreated, models.SuccessMessage{Message: "User registered successfully"})
}

// @Summary Login user
// @Description Authenticate user and return JWT token
// @Tags Auth
// @Accept json
// @Produce json
// @Param request body models.LoginRequest true "Login info"
// @Success 200 {object} models.SuccessMessage
// @Failure 400 {object} models.ErrorResponse
// @Failure 401 {object} models.ErrorResponse
// @Router /auth/login [post]
func (h *AuthHandler) Login(c *gin.Context) {
	var req models.LoginRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, models.ErrorResponse{Error: "Invalid input"})
		return
	}

	id, passwordHash, role, err := h.ar.GetUserByEmail(req.Email)
	if err != nil {
		c.JSON(http.StatusUnauthorized, models.ErrorResponse{Error: "Invalid email or password"})
		return
	}

	if err := bcrypt.CompareHashAndPassword([]byte(passwordHash), []byte(req.Password)); err != nil {
		c.JSON(http.StatusUnauthorized, models.ErrorResponse{Error: "Invalid email or password"})
		return
	}

	token, err := middleware.GenerateJWT(id, role)
	if err != nil {
		c.JSON(http.StatusInternalServerError, models.ErrorResponse{Error: "Failed to generate token"})
		return
	}

	c.JSON(http.StatusOK, gin.H{"message": "Login successful", "token": token})
}

// // @Summary Get user profile
// // @Description Get profile information of logged-in user
// // @Tags Auth
// // @Produce json
// // @Success 200
// // @Failure 401 {object} models.ErrorResponse
// // @Security ApiKeyAuth
// // @Router /auth/profile [get]
// func (h *AuthHandler) Profile(c *gin.Context) {
// 	userID, _ := c.Get("userID")
// 	role, _ := c.Get("role")

// 	c.JSON(http.StatusOK, gin.H{
// 		"userID": userID,
// 		"role":   role,
// 	})
// }

// Logout godoc
// @Summary      Logout user
// @Description  Mem-blacklist JWT token agar tidak bisa dipakai lagi.
// @Tags         Auth
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Success      200  {object} map[string]string
// @Failure      400  {object} map[string]string
// @Failure      401  {object} map[string]string
// @Router       /auth/logout [post]
func (h *AuthHandler) Logout(c *gin.Context) {

	tokenString := c.GetString("token")
	if tokenString != "" {
		err := h.RedisClient.Set(c, "blacklist:"+tokenString, "true", time.Hour*24).Err()
		if err != nil {
			c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to blacklist token"})
			return
		}
	}

	c.JSON(http.StatusOK, gin.H{
		"message": "Logout berhasil",
	})
}